requirements.txt
django
django-debug-toolbar>=1.8
jsonplus
six
sqlalchemy
